requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
Flask==3.1.2
Flask-Cors==6.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==26.0
psycopg2-binary==2.9.11
pycparser==2.23
PyMySQL==1.1.2
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
